:pencil2: (Singup) Rename new signup without V2
