+++ src/features/auth/pages/signup/SetPassword/SetPassword.native.test.tsx
@@ -2,25 +2,25 @@ import React from 'react'
 
 import { act, fireEvent, render, screen } from 'tests/utils'
 
-import { SetPasswordV2 } from './SetPasswordV2'
+import { SetPassword } from './SetPassword'
 
 const props = { goToNextStep: jest.fn(), signUp: jest.fn() }
 
-describe('SetPasswordV2 Page', () => {
+describe('SetPassword Page', () => {
   it('should display security rules', () => {
-    render(<SetPasswordV2 {...props} />)
+    render(<SetPassword {...props} />)
 
     expect(screen.getByText('12 caractères')).toBeTruthy()
   })
 
   it('should disable the submit button when password is incorrect', () => {
-    render(<SetPasswordV2 {...props} />)
+    render(<SetPassword {...props} />)
 
     expect(screen.getByTestId('Continuer')).toBeDisabled()
   })
 
   it('should enable the submit button when password is correct', async () => {
-    render(<SetPasswordV2 {...props} />)
+    render(<SetPassword {...props} />)
 
     const passwordInput = screen.getByPlaceholderText('Ton mot de passe')
     fireEvent.changeText(passwordInput, 'user@AZERTY123')
@@ -29,7 +29,7 @@ describe('SetPasswordV2 Page', () => {
   })
 
   it('should go to next step when submitting password', async () => {
-    render(<SetPasswordV2 {...props} />)
+    render(<SetPassword {...props} />)
 
     await act(async () => {
       const passwordInput = screen.getByPlaceholderText('Ton mot de passe')
