+++ src/features/navigation/RootNavigator/routes.ts
@@ -8,8 +8,8 @@ import { AfterSignupEmailValidationBuffer } from 'features/auth/pages/signup/Aft
 import { NotYetUnderageEligibility } from 'features/auth/pages/signup/NotYetUnderageEligibility/NotYetUnderageEligibility'
 import { SignupConfirmationEmailSent } from 'features/auth/pages/signup/SignupConfirmationEmailSent/SignupConfirmationEmailSent.deprecated'
 import { SignupConfirmationExpiredLink } from 'features/auth/pages/signup/SignupConfirmationExpiredLink/SignupConfirmationExpiredLink'
+import { SignupForm as NewSignupForm } from 'features/auth/pages/signup/SignupForm'
 import { SignupFormDeprecated } from 'features/auth/pages/signup/SignupForm.deprecated'
-import { SignupForm as NewSignupForm } from 'features/auth/pages/signup/SignupFormV2'
 import { VerifyEligibility } from 'features/auth/pages/signup/VerifyEligiblity/VerifyEligibility'
 import { AccountReactivationSuccess } from 'features/auth/pages/suspendedAccount/AccountReactivationSuccess/AccountReactivationSuccess'
 import { FraudulentAccount } from 'features/auth/pages/suspendedAccount/FraudulentAccount/FraudulentAccount'
